requirements.txt
flask
deepface
opencv-python-headless
mtcnn
pillow
pdf2image
numpy
mediapipe
streamlit
PyMuPDF